Update lyrics command

+++ src/commands/music/lyrics.ts
@@ -22,10 +22,17 @@ export const command: App.Command = {
 					: null;
 
 		try {
-			const results = await player.lyrics.search({
-				q: query ?? `${queue?.currentTrack?.title}`,
-			});
-			const lyrics = results?.[0];
+			const results = await player.lyrics.search(
+				query
+					? {
+							q: query,
+						}
+					: {
+							trackName: queue?.currentTrack?.cleanTitle,
+							artistName: queue?.currentTrack?.author,
+						}
+			);
+			const lyrics = results[0];
 			const trimmedLyrics = lyrics.plainLyrics.substring(0, 1997);
 			const embed = new EmbedBuilder()
 				.setTitle(lyrics.trackName)
@@ -36,29 +43,25 @@ export const command: App.Command = {
 			if (!lyrics?.plainLyrics)
 				return await App.respond(command, `❌ | There are no available lyrics for this track`);
 
-			if (!query) {
-				try {
-					const syncedLyrics = queue?.syncedLyrics(lyrics);
+			if (!query && lyrics.syncedLyrics) {
+				const syncedLyrics = queue?.syncedLyrics(lyrics);
 
-					if (syncedLyrics?.isSubscribed()) {
-						syncedLyrics.pause();
-						syncedLyrics.unsubscribe();
+				if (syncedLyrics?.isSubscribed()) {
+					syncedLyrics.pause();
+					syncedLyrics.unsubscribe();
 
-						return await App.respond(command, '❎ | Stopped lyrics');
-					} else {
-						syncedLyrics?.onChange(async (lyrics, timestamp) => {
-							const response = await App.respond(command, `**${Util.formatDuration(timestamp)}**: ${lyrics}`, {
-								channelSend: true,
-							});
-
-							setTimeout(() => response.delete(), 7_000);
+					return await App.respond(command, '❎ | Stopped lyrics');
+				} else {
+					syncedLyrics?.onChange(async (lyrics, timestamp) => {
+						const response = await App.respond(command, `**${Util.formatDuration(timestamp)}**: ${lyrics}`, {
+							channelSend: true,
 						});
-						syncedLyrics?.subscribe();
 
-						return await App.respond(command, '🔄️ | Syncing lyrics');
-					}
-				} catch {
-					// Do nothing, continue.
+						setTimeout(() => response.delete(), 7_000);
+					});
+					syncedLyrics?.subscribe();
+
+					return await App.respond(command, '🔄️ | Syncing lyrics');
 				}
 			}
